requirements.txt
fastapi
uvicorn
python-multipart
torch==2.0.1+cpu
torchvision==0.15.2+cpu
-f https://download.pytorch.org/whl/cpu/torch_stable.html
opencv-python
dlib
Pillow
imageio
scipy
numpy

ffmpeg-python==0.2.0
imageio-ffmpeg==0.4.7
matplotlib==3.6.0
pandas==1.5.0
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
scikit-image==0.19.3
scikit-learn==1.1.2
tqdm==4.64.1
slowapi
python-jose[cryptography]
passlib[bcrypt]
boto3
requests
python-dotenv

